requirements.txt
Django~=5.1.5
django-allauth~=65.3.1
django-ckeditor==6.5.1
crispy-bootstrap5~=2024.10
django-cleanup
django-cors-headers
django-crispy-forms~=2.3
django-jazzmin
django-debug-toolbar~=5.0.1
django-environ~=0.12.0
django-js-asset
django-redis
django-smart-selects~=1.7.2
django-static-jquery-ui
django-tabbed-changeform-admin
django-uuslug
django-watson
mysqlclient
Pillow~=11.1.0
redis
requests
sorl-thumbnail~=12.11.0
Wand
zipp
